aemfvp-a: Add changes to support busybox and distro boot on Armv-A Base AEM FVP
Add changes in the existing build scripts to support busybox and
distribution boot of Armv-A Base AEM FVP Platform.
The following boot methods are verified by using these changes,
1. AARCH64 :  UEFI + busybox
2. AARCH64 :  U-Boot + busybox
3. AARCH64 :  Distribution (Ubuntu, Fedora, and Debian) boot

And the following changes are also included in this commit.

1. Remove the dependency on uefi-tools. These changes are merged
   from 'refinfra/fb1917c0'.
2. Use the latest GNU v10.3 toolchain from developer.arm.com.
3. Add build-test-busybox.sh for base FVP to create the disk
   image for the busybox boot.
4. Add the fvp platform name in build-test-uefi.sh for distribution
   boot support.

+++ build-busybox.sh
@@ -73,9 +73,11 @@ do_clean ()
 		mkdir -p $BUSYBOX_OUT_DIR
 		make O=$BUSYBOX_OUT_DIR clean
 		popd
-		pushd $TOP_DIR/$BUSYBOX_RAMDISK_PATH
-		rm -f ramdisk.img busybox
-		popd
+		if [ "$PLATFORM" != "aemfvp-a" ]; then
+			pushd $TOP_DIR/$BUSYBOX_RAMDISK_PATH
+			rm -f ramdisk.img busybox
+			popd
+		fi
 	fi
 }
 
@@ -83,29 +85,31 @@ do_package ()
 {
 	if [ "$BUSYBOX_BUILD_ENABLED" == "1" ]; then
 		echo "Packaging BUSYBOX... $VARIANT";
-		# create the ramdisk
-		pushd $TOP_DIR/$BUSYBOX_RAMDISK_PATH
-		cp $TOP_DIR/$BUSYBOX_RAMDISK_BUSYBOX_PATH .
-		$TOP_DIR/$LINUX_PATH/$LINUX_OUT_DIR/$LINUX_CONFIG_DEFAULT/usr/gen_init_cpio files.txt > ramdisk.img
-		popd
-		# Copy binary to output folder
-		pushd $TOP_DIR
-		mkdir -p ${OUTDIR}
-		cp $BUSYBOX_RAMDISK_PATH/ramdisk.img \
-			${PLATDIR}/ramdisk-busybox.img
-		popd
-		if [ "$UBOOT_BUILD_ENABLED" == "1" ]; then
-			pushd ${PLATDIR}
-			for target in $TARGET_BINS_PLATS; do
-				local addr=TARGET_$target[ramdisk]
-				${UBOOT_MKIMG} -A $BUSYBOX_ARCH -O linux -C none \
-					-T ramdisk -n ramdisk \
-					-a ${!addr} -e ${!addr} \
-					-n "BusyBox ramdisk" \
-					-d ramdisk-busybox.img \
-					uInitrd-busybox.${!addr}
-			done
+		if [ "$PLATFORM" != "aemfvp-a" ]; then
+			# create the ramdisk
+			pushd $TOP_DIR/$BUSYBOX_RAMDISK_PATH
+			cp $TOP_DIR/$BUSYBOX_RAMDISK_BUSYBOX_PATH .
+			$TOP_DIR/$LINUX_PATH/$LINUX_OUT_DIR/$LINUX_CONFIG_DEFAULT/usr/gen_init_cpio files.txt > ramdisk.img
 			popd
+			# Copy binary to the output folder
+			pushd $TOP_DIR
+			mkdir -p ${OUTDIR}
+			cp $BUSYBOX_RAMDISK_PATH/ramdisk.img \
+				${PLATDIR}/ramdisk-busybox.img
+			popd
+			if [ "$UBOOT_BUILD_ENABLED" == "1" ]; then
+				pushd ${PLATDIR}
+				for target in $TARGET_BINS_PLATS; do
+					local addr=TARGET_$target[ramdisk]
+					${UBOOT_MKIMG} -A $BUSYBOX_ARCH -O linux -C none \
+						-T ramdisk -n ramdisk \
+						-a ${!addr} -e ${!addr} \
+						-n "BusyBox ramdisk" \
+						-d ramdisk-busybox.img \
+						uInitrd-busybox.${!addr}
+				done
+				popd
+			fi
 		fi
 	fi
 }
